refactor: simplify human task backend
